Created template of genetic algorithm behaviour.
Implements gen_server, and stores a two-dimensional grid of processes
that represent individuals.

Handles neighbor events by sending a list back to the process of all
of its neighboring processes.

Probably need to more ga record definition into header file, so that
implementations can use the state to decide when to terminate.