Defined callback should_terminate/3 for stopping criterion.
The function is given (1) the fitness of the individual, (2) the
generation number, and (3) the age. Using some combination of these
parameters, the process should make a decision of whether or not to
terminate.

If the process does decide to terminate, then upon receive a message
to evolve, it returns {stop, normal, State}. The genetic algorithm
process monitors each individual process, and will receive a 'DOWN'
message. The remaining individuals are then stopped as well.

This implementation could probably be improved by using a supervisor
hierarchy. There could be something to find in restarting a process
with an initial state in a world that has already partially converged
to an optimum.